requirements.txt
Flask==2.3.3
Werkzeug==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Login>=0.6.3
Flask-WTF==1.1.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0
email-validator==2.0.0
scipy>=1.11.1
numpy>=1.24.3
gunicorn==21.2.0
pytest>=7.0.0
pytest-flask>=1.2.0
pytest-cov>=4.0.0
factory-boy>=3.0.0
Faker>=18.0.0
flask_migrate>=4.1.0